requirements.txt
Django==3.1.3
django-extensions==3.1.0
djangorestframework==3.12.2
facebook-sdk==3.1.0
requests==2.25.0